requirements.txt
streamlit>=1.35.0
groq>=0.8.0
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pypdf>=4.0.0
pydantic>=2.0.0
